// api/src/game/utils/models/enums/planet-composition-type.ts
export enum PlanetCompositionType {
  CHTHONIAN = 'Chthonian planet',
  CARBON = 'Carbon planet',
  CORELESS = 'Coreless planet',
  DESERT = 'Desert planet',
  GAS_DWARF = 'Gas dwarf',
  GAS_GIANT = 'Gas giant',
  HELIUM = 'Helium planet',
  HYCEAN = 'Hycean planet',
  ICE_GIANT = 'Ice giant',
  ICE = 'Ice planet',
  IRON = 'Iron planet',
  LAVA = 'Lava planet',
  OCEAN = 'Ocean planet',
  PROTOPLANET = 'Protoplanet',
  PUFFY = 'Puffy planet',
  SILICATE = 'Silicate planet',
  TERRESTRIAL = 'Terrestrial planet',
}